requirements.txt
nltk
torch
opencv-python
scipy
numpy
pillow
SpeechRecognition